==> scripts/01-download_data.R <==
#### Preamble ####
# Purpose: Dataset Downloading for research
# Author: Tianen (Evan) Hao
# Date: 11 March 2024 
# Contact: evan.hao@mail.utoronto.ca 
# License: MIT
# Pre-requisites: none


# The download of the raw data requires filtering and browsing The General Social Survey (https://gss.norc.org/documents/stata/GSS_stata.zip). 
# The whole replication package is available on-site for downloading regarding to the GSS Codebook.

# Refer back to Readme for instructions downloading the data set for replication.
